Chg: [U済] CPropCommonクラス整理(CPropComKeyBind) Patches: #480
git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk@2469 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/CKeyBind.h
@@ -33,7 +33,7 @@ struct KEYDATA {
 		SHIFT, CTRL, ALTの３つのシフト状態のそれぞれに対して
 		機能を割り当てるため、配列になっている。
 	*/
-	int	m_nFuncCodeArr[8];
+	short	m_nFuncCodeArr[8]; // 2012.11.8 aroka KEYDATAINITとサイズを合わせる
 };
 
 class CFuncLookup;
